+++ jovo-model/src/index.ts
@@ -1,5 +1,10 @@
+import * as JovoModelValidator from './validators/JovoModelValidator.json';
+import * as JovoModelValidatorV3 from './validators/JovoModelValidatorV3.json';
+
 export * from './Interfaces';
 export * from './JovoModelHelper';
 export * from './JovoModelBuilder';
 export * from './JovoModel';
 export * from './ModelValidationError';
+
+export { JovoModelValidator, JovoModelValidatorV3 };