+++ packages/jovo-model-google/src/index.ts
@@ -3,6 +3,10 @@ import {GoogleActionLanguageModelProperty} from './Interfaces';
 export * from './JovoModelGoogle';
 
 declare module '@jovotech/model' {
+  interface InvocationObject {
+    googleAssistant: string;
+  }
+
   interface JovoModelData {
     googleAssistant?: {
       custom?: {